implemented expandable blocks (tests not implemented)
used the preexisting expandable? property to check if a block is expandable, and if it is, increase minAmount by 1. Currently works when you increase/decrease minAmount in main.tsx and change the expandable propertry (e.g it demonstrates the expected output). Not sure if it will still work with a drag and drop library.